+++ src/main/java/Domain/Game/GameRecord.java
@@ -1,7 +1,5 @@
 package Domain.Game;
 
-import Domain.Board.Move;
-
 import java.time.LocalDateTime;
 import java.util.List;
