+++ src/main/java/Domain/Game/MoveHistory.java
@@ -1,7 +1,5 @@
 package Domain.Game;
 
-import Domain.Board.Move;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
